externs: settle six more cross-TU disagreements about a shared object
A tree-wide comparison of every spelled type against every other spelling of
the same symbol finds sixteen disagreements.  Six of them are settleable, and
in each case one side is simply wrong:

  __float_nan / __float_huge are `unsigned long[]` where float.c defines them
  and where MSL's own float.h declares them; math.h and dolphin/math.h said
  `int[]`.
  _stack_end / _stack_addr are `char[]` in __ppc_eabi_linker.h, which is where
  linker-script symbols belong; OSThread.c kept its own `u8[]` copies with a
  comment saying where they really come from -- it now includes that header.
  sReverbStdDelayLengths is NOT const: making the definition const moves it out
  of .data and costs 16 bytes of matched data, so the reader's `const` goes
  instead, and the declaration moves into __axfx.h.
  gMinimapAreaNameAlpha is spelled `short` at its definition and `s16` in
  pause_menu_api.h.

Plus four redundant declarations deleted where a header the file already
includes says exactly the same thing (gAssetLoadCompletedFlags,
gForceLoadImmediately, gRomListLoadInFlight in objprint_dolphin.c;
gMapBlockCellEntryTables in shader.c).

Measured: all 1013 source objects md5-IDENTICAL to the parent build.  Tree
99.81564 / matched_data 1203317 / 906 of 1050; score_delta_gate VERDICT CLEAN
(0 REGRESSED / 0 IMPROVED / 0 LOST / 0 NEW / 0 NEW-UNSCORED / 0 DEMOTED);
ninja all_source EXIT=0; forced main.dol: OK, sha1
e750e8e894707a52446118a4b84f1b58b677b269; banned_shapes 102/83/19 unchanged.

+++ include/math.h
@@ -7,8 +7,8 @@
 #define _MATH_INLINE static inline
 #endif
 
-extern int __float_nan[];
-extern int __float_huge[];
+extern unsigned long __float_nan[];
+extern unsigned long __float_huge[];
 extern int __double_huge[];
 
 #define INFINITY (*(float*)__float_huge)
